Fixed Server Client
Made Slave a dumb updater it only sends update when requested to. Moved
the listening part of the slave into the client frame. The client Frame
now runs the thread and handles the socket this means network handler
will no longer be doing anything with the client part just setting up
the master connections